Refactored parsing directory so that new parsers can be added and registered to specific file type. Added docx document for testing. Started working on docx parser.
