+++ pkg/www/retext/vue.config.js
@@ -2,4 +2,4 @@ module.exports = {
     devServer: {
         proxy: 'http://localhost:3000'
     }
-}
+};
